🎨 Prettify tremendous doc updates
Great work! The clear structure and detailed sections, like "How to Call `simulateTransaction`" and "Understanding the Footprint" give a thorough framework. 👍🏻 I particularly appreciate the practical examples using both Fetch and the JavaScript SDK, which make it easy for implementation easy. 🧠 The breakdown of `SimulateTransactionParams` and `SimulateTransactionResult` is very well-explained, ensuring readers grasp the intricacies of these interfaces. 👀 Lastly, the thoughtful inclusion of common preflight errors and how to handle them shows a deep understanding of needs and potential challenges.